make update & remove & add a Product

+++ api_coffeeStore/src/main/java/com/example/api_coffeestore/repository/CategoryRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepo extends JpaRepository<Category,Long> {
+public interface CategoryRepo extends JpaRepository<Category,Integer> {
     @Query(value = "SELECT b FROM Category b WHERE b.name IN :name")
     Category findCategoryByName(@Param("name") String name);
 }
